move delocalizedTriplets code to unit.resonance
